requirements.txt
streamlit>=1.30.0
langchain>=0.1.0
langchain-anthropic>=0.1.23
anthropic>=0.34.0
python-dotenv>=1.0.0
pytest>=7.4.0
